api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.0.1
langchain==0.0.335
llama-cpp-python==0.2.19
paddleocr==2.7.0
layoutlmft==0.1.1
transformers==4.34.1
torch==2.1.0
numpy==1.26.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
camelot-py==0.11.0
tabula-py==2.8.2
redis==5.0.1
httpx==0.25.0
# New dependencies for enhanced document processing
unstructured==0.10.30
unstructured-inference==0.7.19
marker-pdf==0.1.5
